+++ Sources/BugsnagPerformance/Private/SpanAttributesProvider.h
@@ -18,5 +18,6 @@ public:
     NSDictionary *appStartSpanAttributes(NSString *firstViewName, bool isColdLaunch) noexcept;
     NSDictionary *appStartPhaseSpanAttributes(NSString *phase) noexcept;
     NSDictionary *viewLoadSpanAttributes(NSString *className, BugsnagPerformanceViewType viewType) noexcept;
+    NSDictionary *viewLoadPhaseSpanAttributes(NSString *className, NSString *phase) noexcept;
 };
 }